ektp-python/requirements.txt
google-cloud-storage==1.16.1
google-cloud-vision==0.38.0
